requirements.txt
flask==3.1.0
line-bot-sdk==3.14.0
apscheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1
gunicorn==23.0.0
